COMP: Revert "BUG: LibXml2 is required for building."
Commit r22008 explicitly disable the search of library not needed to
build LibArchive. Then, commit r22009 has been pushed to re-enable the
lookup of LibXml2.

The original intent of commit r22009 was to fix
a build error occurring when configuring LibArchive in an existing
build tree with option -DCMAKE_DISABLE_FIND_PACKAGE_LibXml2:BOOL=OFF.

Since it is required to delete the LibArchive build tree so that the
variable "HAVE_LIBXML_XMLREADER_H" is not set by a previous lookup
of LibXml2, this new commit is reverted.

Forcing the value "HAVE_LIBXML_XMLREADER_H" to a "notfound" value
will also ensure that re-build of existing tree won't cause problem.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@22012 3bd1e089-480b-0410-8dfb-8563597acbee